refactor: DOM 조작 함수들 DOMUtils로 추출
